# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "EntityCreate": {
        "description": "Schema for creating a new entity.",
        "properties": {
          "accounting_record_id": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accounting Record Id"
          },
          "business_address": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Address"
          },
          "business_city": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business City"
          },
          "business_state": {
            "anyOf": [
              {
                "maxLength": 2,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business State"
          },
          "business_zip": {
            "anyOf": [
              {
                "maxLength": 10,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Zip"
          },
          "dba_names": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dba Names"
          },
          "ein": {
            "anyOf": [
              {
                "maxLength": 20,
                "pattern": "^\\d{2}-\\d{7}$|^$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ein"
          },
          "entity_type": {
            "$ref": "#/components/schemas/EntityType",
            "default": "llc"
          },
          "fiscal_year_end": {
            "anyOf": [
              {
                "pattern": "^(0[1-9]|1[0-2])-(0[1-9]|[12]\\d|3[01])$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fiscal Year End"
          },
          "formation_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formation Date"
          },
          "formation_state": {
            "anyOf": [
              {
                "maxLength": 2,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formation State"
          },
          "is_active": {
            "default": true,
            "title": "Is Active",
            "type": "boolean"
          },
          "legal_name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Legal Name",
            "type": "string"
          },
          "mailing_address": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mailing Address"
          },
          "mailing_city": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mailing City"
          },
          "mailing_state": {
            "anyOf": [
              {
                "maxLength": 2,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mailing State"
          },
          "mailing_zip": {
            "anyOf": [
              {
                "maxLength": 10,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mailing Zip"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "purpose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purpose"
          }
        },
        "required": [
          "legal_name"
        ],
        "title": "EntityCreate",
        "type": "object"
      },
      "EntityListResponse": {
        "description": "Schema for paginated entity list response.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EntityResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "EntityListResponse",
        "type": "object"
      },
      "EntityResponse": {
        "description": "Schema for entity response.",
        "properties": {
          "accounting_record_id": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accounting Record Id"
          },
          "business_address": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Address"
          },
          "business_city": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business City"
          },
          "business_state": {
            "anyOf": [
              {
                "maxLength": 2,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business State"
          },
          "business_zip": {
            "anyOf": [
              {
                "maxLength": 10,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Zip"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "dba_names": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dba Names"
          },
          "deleted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "deleted_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted By"
          },
          "ein": {
            "anyOf": [
              {
                "maxLength": 20,
                "pattern": "^\\d{2}-\\d{7}$|^$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ein"
          },
          "entity_type": {
            "$ref": "#/components/schemas/EntityType",
            "default": "llc"
          },
          "fiscal_year_end": {
            "anyOf": [
              {
                "pattern": "^(0[1-9]|1[0-2])-(0[1-9]|[12]\\d|3[01])$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fiscal Year End"
          },
          "formation_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formation Date"
          },
          "formation_state": {
            "anyOf": [
              {
                "maxLength": 2,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formation State"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "default": true,
            "title": "Is Active",
            "type": "boolean"
          },
          "legal_name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Legal Name",
            "type": "string"
          },
          "mailing_address": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mailing Address"
          },
          "mailing_city": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mailing City"
          },
          "mailing_state": {
            "anyOf": [
              {
                "maxLength": 2,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mailing State"
          },
          "mailing_zip": {
            "anyOf": [
              {
                "maxLength": 10,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mailing Zip"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "purpose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purpose"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          }
        },
        "required": [
          "legal_name",
          "created_at",
          "updated_at",
          "id"
        ],
        "title": "EntityResponse",
        "type": "object"
      },
...
      "EntityUpdate": {
        "description": "Schema for updating an existing entity.",
        "properties": {
          "accounting_record_id": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accounting Record Id"
          },
          "business_address": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Address"
          },
          "business_city": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business City"
          },
          "business_state": {
            "anyOf": [
              {
                "maxLength": 2,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business State"
          },
          "business_zip": {
            "anyOf": [
              {
                "maxLength": 10,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Zip"
          },
          "dba_names": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dba Names"
          },
          "ein": {
            "anyOf": [
              {
                "maxLength": 20,
                "pattern": "^\\d{2}-\\d{7}$|^$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ein"
          },
          "entity_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EntityType"
              },
              {
                "type": "null"
              }
            ]
          },
          "fiscal_year_end": {
            "anyOf": [
              {
                "pattern": "^(0[1-9]|1[0-2])-(0[1-9]|[12]\\d|3[01])$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fiscal Year End"
          },
          "formation_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formation Date"
          },
          "formation_state": {
            "anyOf": [
              {
                "maxLength": 2,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formation State"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "legal_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Name"
          },
          "mailing_address": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mailing Address"
          },
          "mailing_city": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mailing City"
          },
          "mailing_state": {
            "anyOf": [
              {
                "maxLength": 2,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mailing State"
          },
          "mailing_zip": {
            "anyOf": [
              {
                "maxLength": 10,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mailing Zip"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "purpose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purpose"
          }
        },
        "title": "EntityUpdate",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "email": "byron@williamscpa.com",
      "name": "Byron Williams",
      "url": "https://github.com/ByronWilliamsCPA/llc-manager"
    },
    "description": "REST API for managing LLC entities, ownership structures, compliance dates, bank accounts, tax filings, and associated documentation.\n\nAll resource endpoints are namespaced under `/api/v1/` to match the API version `1.0.0`. Health probe endpoints are exposed under `/api/health/` and follow Kubernetes probe conventions.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/ByronWilliamsCPA/llc-manager/blob/main/LICENSE"
    },
    "title": "LLC Manager API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/entities": {
      "get": {
        "description": "Return a paginated, optionally filtered list of LLC entities. Supports full-text search across `legal_name`, `ein`, and `dba_names`, plus an `is_active` filter. Soft-deleted entities are excluded.",
        "operationId": "list_entities_api_v1_entities_get",
        "parameters": [
          {
            "description": "Page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Page number",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page",
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Items per page",
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "description": "Search by legal name or EIN",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search by legal name or EIN",
              "title": "Search"
            }
          },
          {
            "description": "Filter by active status",
            "in": "query",
            "name": "is_active",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by active status",
              "title": "Is Active"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityListResponse"
                }
              }
            },
            "description": "Paginated list of entities"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List entities",
        "tags": [
          "Entities"
        ]
      },
      "post": {
        "description": "Create a new LLC entity. The optional `ein` field must be unique across non-deleted entities; a duplicate value returns 409.",
        "operationId": "create_entity_api_v1_entities_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityResponse"
                }
              }
            },
            "description": "Entity created successfully"
          },
          "409": {
            "description": "Entity with the supplied EIN already exists"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "summary": "Create entity",
        "tags": [
          "Entities"
        ]
      }
    },
    "/api/v1/entities/{entity_id}": {
      "delete": {
        "description": "Soft-delete an entity by stamping `deleted_at`. The record is retained for audit but excluded from list and detail responses.",
        "operationId": "delete_entity_api_v1_entities__entity_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Entity soft-deleted"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Soft delete entity",
        "tags": [
          "Entities"
        ]
      },
      "get": {
        "description": "Fetch a single entity by its UUID. Soft-deleted entities are not returned.",
        "operationId": "get_entity_api_v1_entities__entity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityResponse"
                }
              }
            },
            "description": "Entity retrieved"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get entity by ID",
        "tags": [
          "Entities"
        ]
      },
      "patch": {
        "description": "Partially update an entity. Only fields included in the request body are modified. Updating `ein` to a value already held by another entity returns 409.",
        "operationId": "update_entity_api_v1_entities__entity_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityResponse"
                }
              }
            },
            "description": "Entity updated"
          },
          "404": {
            "description": "Entity not found"
          },
          "409": {
            "description": "Entity with the supplied EIN already exists"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "summary": "Update entity",
        "tags": [
          "Entities"
        ]
      }
    }
  }
}